compare

+++ crl/envelope/models/linear.py
@@ -62,7 +62,7 @@ class EnvelopeLinearCQN(torch.nn.Module):
 
 
 	def forward(self, state, preference, w_num=1):
-		s_num = preference.size(0) / w_num
+		s_num = int(preference.size(0) / w_num)
 		x = torch.cat((state, preference), dim=1)
 		x = x.view(x.size(0), -1)
 		x = F.relu(self.affine1(x))
